FEAT: Apply new template
on startbootstrap
